==> uplift_and_kalman/3_do_preproc.ipynb <==
import pandas as pd
import numpy as np
import tqdm
import scipy.stats as sts

features = pd.read_csv('./large_data/output_data/final_dfdiouuy.csv')
clients = pd.read_csv('./large_data/raw_data/clients.csv', index_col=0)

poor = pd.read_csv('./large_data/raw_data/poor_raw.csv')
rich = pd.read_csv('./large_data/raw_data/rich_raw.csv')

total_df = pd.concat([poor, rich])

del poor
del rich

total_df = total_df[[
    'client_id', 'transaction_id',
    'transaction_datetime', 'purchase_sum',
    'age', 'gender',
    'treatment_flg', 'city_type',
]].drop_duplicates()

total_df['transaction_datetime'] = pd.to_datetime(total_df['transaction_datetime']).dt.date

**разбиваем фичу с полом**

total_df['female'] = (total_df.gender == 'F').astype(int)
total_df['male'] = (total_df.gender == 'M').astype(int)

total_df.drop(columns=['gender'], inplace=True)

**процессинг тагрета**

def modify_products(row):
    ''' Функция для зашумления людей из богатого города, на которых разослали коммуникацию'''
    
    if (row['treatment_flg'] == 1) & (row['city_type'] == 1):
        # Генерируем экспоненциальную случайную величину с матожиданием row['age']
        exp_value = np.random.exponential(5*abs(row['age']))
        return row['purchase_sum'] + exp_value
    return row['purchase_sum']

# Применяем функцию по строкам DataFrame
total_df['purchase_sum'] = total_df.apply(modify_products, axis=1)

total_df = total_df.groupby(['client_id', 'transaction_datetime'], as_index=False).agg({
    'purchase_sum':'sum',
    'treatment_flg':'mean',
    'city_type':'mean',
    'female':'mean',
    'male':'mean',
})
total_df.rename(columns={'transaction_datetime':'transaction_date'}, inplace=True)

total_df.info()

res_data = pd.concat([features, total_df.drop(columns=['client_id','client_id'])],axis=1)

**собираем красивую панель**

res_data = res_data[[
    'client_id','transaction_date',
    'city_type','treatment_flg',
    'age', 'female', 'male',
    'purchase_sum','avg_product_quantity', 
    'total_products', 'ch_avg_purchase',
    'unique_segments', 'unique_brands', 
    'alcohol_percentage','own_trademark_percentage',
]]

res_data.to_csv('./large_data/intermediate_data/panel_data.csv')
